[refactor] Refactored to eliminate the feature envy and code duplication code smells
1. Code Duplication - Created the abstract class Job. This abstract class
gets extended by PolishingJob and CleaningJob. There is a single
method to calculate the rent of a machine for the floor. Having a single
method to calculate rent fixes the code smell of code duplication, as
previously code was being duplicated in CleaningJob and PolishingJob's
rent methods, but now only one method is needed for both types of
machines. In order for this to work, an abstract class Machine was created.
This abstract class gets extended by PolishingMachine and
CleaningMachine. This allows the Job class to take Machine as a parameter
when calculating the rent of a machine, meaning only a single method is
needed for calculating rent for both types of machines. This reduces code
duplication, as code was previously being duplicated in PolishingJob and
CleaningJob's rent methods

2. Feature Envy - Modified the Floor class to include a method called
getArea. The method getArea has been added to fix the code smell of feature
envy. Previously, CleaningJob and PolishingJob were using the length and
width of a Floor object when calculating the floor area, rather than
having the Floor object calculate and return the area. These other classes
using the length and width methods of the Floor class to perform calculations
is feature envy, and now allowing the Floor object to perform the area
calculation means this feature envy is no longer present. In Job's
getRent method, the Floor object now calculates the floor area within its
own class, rather than job classes calling length and width in order to
calculate area. This eliminates the previously present feature envy